requirements.txt
autopep8==1.4.3
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
future==0.17.1
idna==2.7
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
nbformat==4.4.0
nltk==3.3
numpy==1.15.4
oauthlib==2.1.0
parso==0.3.1
Pillow==5.3.0
plotly==3.3.0
pluggy==0.8.0
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-jsonrpc-server==0.0.2
python-language-server==0.21.3
pytz==2018.7
requests==2.19.1
requests-oauthlib==1.0.0
retrying==1.3.3
rope==0.11.0
six==1.11.0
snowballstemmer==1.2.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.1.0
Werkzeug==0.14.1
yapf==0.25.0
